+++ src/main/java/com/webLibraryManagementSystem/controllers/DeleteMemberServlet.java
@@ -1,8 +1,11 @@
 package com.webLibraryManagementSystem.controllers;
 
 import java.io.IOException;
+import java.util.List;
 
+import com.webLibraryManagementSystem.domain.CustomActiveIssuedBooks;
 import com.webLibraryManagementSystem.domain.Member;
+import com.webLibraryManagementSystem.services.impl.BookServicesImpl;
 import com.webLibraryManagementSystem.services.impl.MemberServiceImpl;
 import com.webLibraryManagementSystemexceptions.InvalidException;
 
@@ -25,6 +28,7 @@ public class DeleteMemberServlet extends HttpServlet {
 	}
 
 	MemberServiceImpl memberService = new MemberServiceImpl();
+	BookServicesImpl bookService = new BookServicesImpl();
 
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -32,8 +36,14 @@ public class DeleteMemberServlet extends HttpServlet {
 		int memberId = Integer.parseInt(request.getParameter("memberId"));
 		try {
 			Member member = memberService.getMemberId(memberId);
-			memberService.deleteMember(member);
-			request.setAttribute("message", member.getName() + "deleted Successfully");
+			List<CustomActiveIssuedBooks> books = bookService.getActiveIssuedBooks();
+			long count = books.stream().filter(isuue -> isuue.getMemberId() == member.getMemberId()).count();
+			if (count > 0) {
+				request.setAttribute("message", "Member have " + count + " Books Please collect books before delete");
+			} else {
+				memberService.deleteMember(member);
+				request.setAttribute("message", member.getName() + "deleted Successfully");
+			}
 
 		} catch (InvalidException e) {
 			request.setAttribute("message", e.getMessage());
